requirements.txt
Django==5.1.4
requests==2.32.3
pyzmq==26.2.0
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2
python-dotenv==1.0.1
